Refer to documentation - installation

+++ docs/src/_build.sh
@@ -1,10 +1,28 @@
 #!/bin/sh
 
-set -ev
+pids=""
+outputs="git pdf_ epub_"
+for output in $outputs ; do
+    Rscript -e "bookdown::render_book('index.Rmd', 'bookdown::${output}book')" &
+    pids="$pids $!"
+done
 
-Rscript -e "bookdown::render_book('index.Rmd', 'bookdown::gitbook')"
-Rscript -e "bookdown::render_book('index.Rmd', 'bookdown::pdf_book')"
-Rscript -e "bookdown::render_book('index.Rmd', 'bookdown::epub_book')"
+errors=""
+for pid in $pids ; do
+    wait $pid
+    if [ $? -ne 0 ] ; then
+        errors="$errors $pid"
+    fi
+done
+
+if ! [ -z $errors ] ; then
+    echo "\n\nlaunched processes:"
+    echo "$outputs with pids $pids"
+    echo "$errors failed"
+    return 1
+fi
+
+echo "done"
 
 rm -r ../libs/ ../0* || echo "nothing to remove"
 rm -r _book/images
